refactor(jwa): move algorithm types into jose-jwa
