+++ jose-jwa/Cargo.toml
@@ -16,6 +16,9 @@ readme = "README.md"
 edition = "2021"
 rust-version = "1.63"
 
+[dependencies]
+serde = { version = "1.0.136", default-features = false, features = ["alloc", "derive"] }
+
 [package.metadata.docs.rs]
 all-features = true
 rustdoc-args = ["--cfg", "docsrs"]
